Should not interfere loading bridgesupport file of frameworks
If use motion.h to load bridgesupport of library, like
```
  app.include 'sqlite3.h'
```
then it interferes loading bridgesupport file of frameworks
which was specfied by "app.frameworks"

```
  app.include 'sqlite3.h'
  app.frameworks << 'AVFoundation'
```
By this configuration, the bridgesupport file of AVFoundation will not be loaded.

Because the bridgesupport file is listed up when @bridgesupport_files is nil in
https://github.com/HipByte/RubyMotion/blob/ef6f52f57443247d7cbaf92c6b898d4eaf9d2740/lib/motion/project/xcode_config.rb#L237-L251
However, @bridgesupport_files was configured by motion.h, it cannot work properly.

So, this patch will configure @bridgesupport_files before build process.